Populate/display primitive workout feed
In order to display the user's workouts, I added a function in the workout controller to populate the user's list of workoutIds into full workouts, then return that array into the workout reducer and pass it into the component. I'm going to now try to just populate the array when the user is fetched, and try to access the workouts that way, because I'd rather have the workouts stay contained in the user and not floating around in the state.