Fix addWorkout bug for non-team users
Before this commit, I was appending the workouts to the user's list in the if statement for users on a team. This was causing a bug where non-team users' workouts weren't being recorded in their profile, and thus weren't showing up in their workout feed. I took these lines of code outside of that if statement, so it will add workouts now for both team and non-team users alike.